+++ src/include/arch/x86/ports.h
@@ -11,6 +11,7 @@ typedef struct {
 } device_init_t;
 
 void outb(uint16_t port, uint8_t value);
+void outw(uint16_t port, uint16_t value);
 uint8_t inb(uint16_t port);
 uint16_t inw(uint16_t port);
 void iowait(void);