+++ src/include/sys/timekeeping.h
@@ -3,7 +3,6 @@
  * timekeeping.h - Functionality for time
  */
 
-
 typedef struct callback_state
 {
     // Data given by register_callback()
@@ -17,16 +16,17 @@ typedef struct callback_state
     int active;
 } callback_state_t;
 
-
 // Struct for a timekeeping device
-
 typedef struct timekeeping_state
 {
     uint32_t frequency;     // Frequency of clock
     uint64_t ticks;         // Simple number of ticks
 
+    uint32_t priority;      // Priority to set system clock (highest wins)
+
     uint32_t us_per_tick;   // Microseconds per tick (rounded)
     uint32_t ns_per_tick;   // Remainder from microsecond round (3 places)
     uint32_t ns_count;      // Count of ns, so we can add it to timestamp
+  
 } timekeeping_state_t;