+++ CHANGELOG.md
@@ -4,6 +4,18 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.3] - 2022-04-40
+### Added
+- ipxe iso image 
+- Kconfig files
+- Modular Makefile system
+- Qemu-specific shutdown driver
+- Shutdown function in HAL
+### Changed
+- Rewrote Multiboot code
+- Modified Code Struture
+- Changed build system entirely
+
 ## [0.0.2] - 2018-12-12
 ### Added
 - MIT License tag for README.md